refactor: 패키지 정리 및 사용하지 않는 메소드 제거
